+++ CHANGELOG.md
@@ -20,6 +20,9 @@ Changelog follow https://keepachangelog.com/ format.
     *   Added: `lazy_imports.print_current_imports` to display the active lazy
         imports (e.g. to add imports before publishing a colab).
 *   `epy`:
+    *   Added: `epy.ContextManager` to create yield-based contextmanager class
+        (see
+        [discussion](https://discuss.python.org/t/yield-based-contextmanager-for-classes/8453))
     *   Added: `epy.issubclass` (like `issubclass` but does not raises error for
         non-types)
     *   Added: `epy.groupby`, like `itertools.groupby` but returns a `dict`.